Remove the socket package from the runtime
The socket package was used for accessing PDG live data. But the
website is now using https and luasocket does not natively support it.
LuaSec would be needed which depends on OpenSSL. Bundling these extra
deps in the runtime is beyond the present scope. If those packages are
needed one should install them manually or use another runtime.